fix: 削除できないステータスのチェックを追加 (#35)

+++ src/usecase/task/deleteTaskUseCase.ts
@@ -2,6 +2,7 @@ import type ITaskRepository from '$/server/task/taskRepository';
 import type TaskId from '$/server/task/valueObject/taskId';
 import type { UserSession } from '^/usecase/shared/userSession';
 import { inject, singleton } from 'tsyringe';
+import InvalidControl from '$/shared/exceptions/domain/invalidControl';
 import Forbidden from '$/shared/exceptions/http/forbidden';
 import { fromEntity } from '^/usecase/task/taskDto';
 
@@ -16,6 +17,10 @@ export default class DeleteTaskUseCase {
     const task = await this.repository.findById(taskId);
     if (!task.userId.equals(userSession.userId)) {
       throw new Forbidden();
+    }
+
+    if (!task.canDelete()) {
+      throw new InvalidControl('削除できないステータスです');
     }
 
     await this.repository.delete(task.taskId);
